third_party/blink/web_tests/fast/shapes/shape-outside-floats/shape-outside-big-box-border-radius-001-expected.txt

The green squares should follow the right side of the blue shape. They should not overlap the shape.

X
X
X
X
X
X
PASS elementRect('a').top is 0
PASS elementRect('a').right is 120
PASS elementRect('b').top is 20
PASS elementRect('c').top is 40
PASS elementRect('d').top is 60
PASS elementRect('e').top is 80
PASS elementRect('f').top is 100
PASS elementRect('f').right is 20
PASS successfullyParsed is true

TEST COMPLETE
